requirements.txt
tensorflow>=2.10.0,<3.0.0
streamlit>=1.28.0
numpy>=1.20.0
pillow>=8.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.12.0
opencv-python>=4.5.0

